Fixed the permissions problem
