+++ frontend/cli-go/internal/app/prepare.go
@@ -200,15 +200,3 @@ func normalizeFilePath(path string, workspaceRoot string, cwd string) (string, b
 	}
 	return absPath, false, nil
 }
-
-func isWithin(base, target string) bool {
-	rel, err := filepath.Rel(base, target)
-	if err != nil {
-		return false
-	}
-	if rel == "." {
-		return true
-	}
-	prefix := ".." + string(os.PathSeparator)
-	return !strings.HasPrefix(rel, prefix) && rel != ".."
-}
